ndt.go: send at least one MSG_RESULT back (#18)
The [specification](https://github.com/ndt-project/ndt/wiki/NDTProtocol)
doesn't explicitly state that the server MUST send to the client
at least one MSG_RESULT message. But, looking into the sequence
diagram that is part of the specification, this is CLEARLY implied.

As such, botticelli was broken because it was violating the protocol.

The proper fix would be to read Web100 variables, compute a number
of metrics and send those back to the client. But, botticelli has
not reached this point of maturity yet.

As a fix, I implemented sending back a single variable formatted as
the NDT code would expect, so that it does not crash, but named using
a clearly nonstandard name such that it's not used.

Bug reported by @nkinkade, thanks!